Refactored pyramid buffer processors
